dev-requirements.txt
pytest-ckan
git+https://github.com/DataShades/ckanext-oidc-pkce.git@master#egg=ckanext-oidc-pkce
